Rename lrlex's builder module to ctbuilder for consistency with lrpar.
